Allow the authorization to continue without an id_token.
Servers implementing OAuth 2.0 Code Authorization Grant Flow were
almost working, as the basic flow is the same as the OpenID Connect
Code Flow.  Once the access_token is returned, the only thing missing
was the id_token.  But if OIDCRemoteUserClaim is customized to take
the the REMOTE_USER from information returned by a UserInfo endpoint,
the id_token is not necessary.

Additional changes were made to allow more flexibility in the UserInfo
endpoint.  The aim was to get this working with Phabricator's built-in
OAuthServer, and this required the access_token to be passed to the
user.whoami API method as a parameter, and dereferencing of the user
info from a "result" object contained within the response.

This has been tested with the Phabricator OAuthServer implementation.
Other servers implementing Oauth 2.0 Code Authorization Grant Flow,
and which provide a suitable API for getting UserInfo as a JSON object
may also work.  I tried to make these changes without breaking any
existing functionality, but have not tested against any OpenID Connect
servers, so it cannot be guaranteed.

Note that the changes relax some checks, so if your server implements
OpenID Connect, it may be safer to use the upstream module.