Dereference OIDCProviderUserInfoResponseSubkey earlier
Previously, this option was used when extracting the REMOTE_USER from the
claims.  As a result, it affected only the user id, and the whole JSON object
was forwarded to the application under a single header, which makes the
claims difficult to use.

Now it is used when first processing the claims from the UserInfo endpoint,
so that all claims are retrieved from the sub-key, and stored in the session
and forwarded to the application as separate items.

When working with Phabricator, this should result in OIDC_CLAIM_primaryEmail,
OIDC_CLAIM_realName, OIDC_CLAIM_image, OIDC_roles headers that can be directly
useful, instead of just an OIDC_CLAIM_result.